feat: Controller Input과 일반 Input 분리
