+++ src/components/Input/Input.tsx
@@ -1,58 +1,31 @@
 'use client';
 
-import clsx from 'classnames';
-import { useController, FieldValues, FieldPath, Control, Controller, useForm } from 'react-hook-form';
+import Image from 'next/image';
 import { inputWrapper, input, Size, clearButton } from './Input.css';
 import { ForwardedRef, forwardRef } from 'react';
-import Image from 'next/image';
 import Typography from '../Typography/Typography';
+import { SchemaKeys } from './schema';
 import { getIconUrl } from '../IconButton/utils/getIconUrl';
 
-type TControl<T extends FieldValues> = {
-	control: Control<T>;
-	name: FieldPath<T>;
-	placeholder: string;
+export type InputProps = {
+	name: SchemaKeys;
+	label?: string;
 	size?: Size;
 } & React.ComponentPropsWithoutRef<'input'>;
 
-const Input = forwardRef(function Input(props: TControl<any>, ref: ForwardedRef<HTMLInputElement>) {
-	const { control, name, placeholder, size } = props;
-	const {
-		field: { value, ...rest },
-		fieldState,
-	} = useController({ name, control });
-	const { setValue } = useForm();
-	const clear = (name: string) => {
-		setValue(name, '', {
-			shouldValidate: true,
-			shouldDirty: true,
-		});
-	};
-
+const Input = forwardRef(function Input(props: InputProps, ref: ForwardedRef<HTMLInputElement>) {
+	const { label, name, placeholder, size, ...rest } = props;
 	return (
-		<div className={inputWrapper}>
-			<Controller
-				control={control}
-				defaultValue=""
-				name={name}
-				render={({ field }) => (
-					<input {...rest} className={input({ size })} placeholder={placeholder} value={field.value} ref={ref} />
-				)}
-			/>
-			<button
-				className={clearButton}
-				disabled={!fieldState.isDirty}
-				type="reset"
-				onClick={() => {
-					console.log('clear', name);
-					clear(name);
-				}}
-			>
+		<div className={inputWrapper} id={`${name}_input`}>
+			{label && (
+				<Typography variant="label3" as="label">
+					{label}
+				</Typography>
+			)}
+			<input name={name} className={input({ size })} placeholder={placeholder} ref={ref} {...rest}></input>
+			<button className={clearButton} type="reset">
 				<Image src={getIconUrl('clear', 'default')} alt={'Icon'} width={24} height={24} />
 			</button>
-			<Typography color="red500" variant="label5" as="p">
-				{fieldState.error?.message}
-			</Typography>
 		</div>
 	);
 });
